requirements.txt
# =====================================
# LiveKit Gemini Voice Agent - MVP
# Минимальный набор для Hugging Face Spaces
# =====================================

# ---- CORE LIVEKIT PACKAGES ----
livekit-agents==1.2.4
livekit-plugins-google==1.2.3
livekit-plugins-silero==1.2.3
livekit-plugins-noise-cancellation==0.1.0
livekit-api==1.0.5
livekit==1.0.12

# ---- GOOGLE AI / GEMINI ----
google-genai==1.28.0
google-auth==2.40.3
google-api-core==2.25.1

# ---- WEB & HTTP (для LiveKit + Health Endpoint) ----
aiohttp==3.12.15
websockets==15.0.1

# ---- MEDIA PROCESSING ----
pillow==11.3.0

# ---- RSS PARSING ----
feedparser==6.0.11

# ---- HTTP CLIENT (для N8N webhook) ----
requests==2.32.3

# ---- DATABASE ----
pymongo==4.10.1
dnspython==2.7.0

# ---- SYSTEM & UTILITIES ----
packaging==25.0
setuptools==80.9.0
python-dotenv==1.0.1

# =====================================
# ВАЖНО: Это МИНИМАЛЬНЫЙ набор для MVP!
# Без N8N, без Tavily, без email tools
# Только LiveKit + Gemini + Hardcoded текст
# =====================================
